implement Request #10829  	Enhance Patch Upload process
git-svn-id: http://svn.php.net/repository/pear/pearweb/trunk@234696 c90b9560-bf6c-de11-be94-00142212c4b1

+++ include/bugs/patchtracker.php
@@ -172,7 +172,10 @@ class Bugs_Patchtracker
             }
             list($id, $fname) = $res;
             $file->setName($fname);
-            $file->setValidExtensions(array('txt', 'patch', 'diff'), 'accept');
+            if ($file->getProp('type') != 'text/plain') {
+                return PEAR::raiseError('Error: uploaded patch file must have text/plain' .
+                    ' MIME type (save as patch.txt)');
+            }
             $tmpfile = $file->moveTo($this->patchDir($bugid, $name));
             if (PEAR::isError($tmpfile)) {
                 $this->_dbh->query('DELETE FROM bugdb_patchtracker
